feat: implemented null check before delete operation

+++ back-end/src/main/java/com/braintrain/backend/roadMaps/RoadMapService.java
@@ -41,6 +41,7 @@ public class RoadMapService {
     }
 
     public void delete(RoadMapMeta roadMapMeta) {
+        if(roadMapMeta == null) return;
         repo.deleteById(roadMapMeta.getRoadMapReferenceId());
         metaRepo.delete(roadMapMeta);
     }
